weddingapp/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.1
cryptography==42.0.5
Django==5.0.4
django-ckeditor==6.7.1
django-filter==24.2
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
idna==3.7
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
